requirements.txt
xlrd==1.1.0
redis==2.10.6
pytz==2018.3
jieba==0.39
Django==1.11.10
Whoosh==2.7.4
django_haystack==2.8.0
